feat: make Dockerfile more general-purpose
This commit changes the Dockerfile so that we can use a bind mount
to mount the local dir, rather than having to copy all the files
into the image at build time.

This should allow us to re-use this docker image for other purposes;
in the near future we can push it to an ECR repo so that devs
don't have to build it locally at all.