feat: Add a prototype async get method
Add an async get method that separates the grpc call from interpreting
the results. This is significantly faster depending on how the method is
used.

Add a temporary test that runs the sync and async get methods in groups
of 50 ten times and prints how long it takes. The async method is ~24x
faster in this test.