fix: explicitly close gRPC channel when cycling idle client (#140)
Prior to this commit we didn't ever have an explicit call to close
the gRPC channel; in cases where we are cycling the client out
due to reaching the idle timeout, this could leave open sockets /
file descriptors behind.  This commit updates the code to explicitly
close the channel when replacing the client.