feat: add configurability for number of gRPC channels (#158)
* feat: add configurability for number of gRPC channels

For users who may end up with over 100 concurrent requests in flight,
we need to be able to support multiple gRPC channels. This commit adds
a configuration setting to allow users to specify the desired number
of channels.

This most likely won't doing anything useful until we add the config
flag to disable persistent TCP connections in the gRPC channel
constructor, because we will just end up re-using the same channel
multiple times.