chore: make order in the sorted set functions more explicit (#235)
change the order argument in the sorted set functions to explicitly take
SORT_ASC and SORT_DESC, and treat other values as ascending. This
replaces the previous logic where SORT_DESC or lower was descending, and
SORT_ASC or higher was ascending. Using built-in constants in a
comparison was confusing.