feat: add flags for toggling inclusivity on fetchByScore scores
Add $inclusiveMin and $inclusiveMax to sortedSetFetchByScore to allow
callers to toggle whether the min and max scores should be inclusive
or exclusive. They default to inclusive.

Add tests for the flags.

Change the behavior for an unrecognized sort order from defaulting to
ascending to returning an error.

Make $order not nullable. There isn't a reason for it to be null and it
has a default argument.