chore: remove the type from sorted set scores so they support ints (#242)
Remove the type that the score arguments use in sorted set methods. When
they were explicitly float, it was an error to pass in an int score.
They are now defined as float|int, but are left untyped in the functions
because union types are unsupported in the oldest versions of PHP we
support.

Update data validation to handle the new potential score types.

Rearrange the arguments of sorted set functions that take a score range
to put the min/max scores and min/max inclusive flags next to each other
instead of interleaving them.